Move list files in workspace to Xcode observer, and support split screen
