+++ app/modules/serviceInterfaces/XcodeObserverServiceInterface/Sources/XcodeObserver+Mock.swift
@@ -33,6 +33,9 @@ public final class MockXcodeObserver: XcodeObserver {
 
   public let mutableStatePublisher: CurrentValueSubject<AXState<XcodeState>, Never>
   public var onGetContent: @Sendable (URL) throws -> String = { _ in throw AppError("Could not read content of file") }
+  public var onListFiles: @Sendable (URL) throws -> ([URL], WorkspaceType) = { _ in
+    throw AppError("Could not list files in workspace")
+  }
 
   public var axNotifications: AnyPublisher<AXNotification, Never> {
     Just(AXNotification.applicationActivated).eraseToAnyPublisher()
@@ -44,6 +47,10 @@ public final class MockXcodeObserver: XcodeObserver {
 
   public func getContent(of file: URL) throws -> String {
     try onGetContent(file)
+  }
+
+  public func listFiles(in workspace: URL) throws -> ([URL], WorkspaceType) {
+    try onListFiles(workspace)
   }
 
 }
