+++ app/modules/services/FileSuggestionService/Sources/DefaultFileSuggestionService+Dependency.swift
@@ -9,13 +9,15 @@ import XcodeObserverServiceInterface
 
 extension BaseProviding where
   Self: ShellServiceProviding,
-  Self: FileManagerProviding
+  Self: FileManagerProviding,
+  Self: XcodeObserverProviding
 {
   public var fileSuggestionService: FileSuggestionService {
     shared {
       DefaultFileSuggestionService(
         fileManager: fileManager,
-        shellService: shellService)
+        shellService: shellService,
+        xcodeObserver: xcodeObserver)
     }
   }
 }
